Add Navigation Tabs for Photo/Form sections, split Form inputs into secondary component
